requirements.txt
# Solar Charging Backend Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.25.0

# MQTT client
paho-mqtt==1.6.1

# Job scheduling
APScheduler==3.10.4

# Configuration
pyyaml==6.0.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
